Add case for limit 0 and fix thread freeze
So before this, if no threads could be created, then we'd call manager.add() in a loop many times, even though a thread will never be available.
Previously, requests were also handled in a blocking way if no thread was available. But this would result in a freeze if more than the thread limit tried to create a connection. So now we just wait for a thread to be available instead, which fixes this issue.